Added flink output function

+++ ipper/flink/main.py
@@ -1,4 +1,5 @@
 import json
+import sys
 
 from argparse import Namespace
 from pathlib import Path
@@ -7,6 +8,8 @@ from ipper.flink.wiki import (
     get_flip_main_page_info,
     get_flip_information
 )
+from ipper.flink.output import render_flink_main_page, FLINK_MAIN_PAGE_TEMPLATE
+from ipper.common.constants import DEFAULT_TEMPLATES_DIR
 
 FLIP_CACHE_FILENAME = "flip_wiki_cache.json"
 
@@ -21,27 +24,50 @@ def setup_flink_parser(top_level_subparsers) -> None:
         dest="flink_subcommand",
     )
     setup_wiki_command(main_subparser)
+    setup_output_command(main_subparser)
 
 
-def process_wiki(args: Namespace):
+def process_wiki(args: Namespace) -> None:
+
+    cache_path: Path = Path(args.cache)
+    cache_path.mkdir(parents=True, exist_ok=True)
+    flip_cache_path: Path = cache_path.joinpath(FLIP_CACHE_FILENAME)
+
+    if flip_cache_path.exists() and not args.overwrite:
+        print(f"Cache file {flip_cache_path} already exists. Add --overwrite to redownload")
+        sys.exit(1)
 
     main_page = get_flip_main_page_info()
     flip_data = get_flip_information(main_page, chunk=args.chunk)
 
-    cache_path: Path = Path(args.cache)
-    cache_path.mkdir(parents=True, exist_ok=args.overwrite)
-    flip_cache_path: Path = cache_path.joinpath(FLIP_CACHE_FILENAME)
-
     with open(flip_cache_path, "w", encoding="utf8") as flip_cache_file:
         json.dump(flip_data, flip_cache_file)
+
+
+def process_output(args: Namespace) -> None:
+
+    wiki_cache_path = Path(args.wiki_cache_file)
+    if not wiki_cache_path.exists():
+        raise AttributeError(f"Wiki Cache file {wiki_cache_path} does not exist")
+
+    with open(wiki_cache_path, "r", encoding="utf8") as wiki_cache_file:
+        wiki_cache_data = json.load(wiki_cache_file)
+
+    render_flink_main_page(
+        wiki_cache_data,
+        args.output_file,
+        args.template_dir,
+        args.template_filename,
+    )
 
 
 def setup_wiki_command(main_subparser):
     """Setup the top level wiki command line option."""
 
     wiki_parser = main_subparser.add_parser(
-        "wiki", help="Command for performing wiki related commands"
+        "wiki", help="Command for performing wiki related tasks"
     )
+    wiki_parser.set_defaults(func=wiki_parser.print_help)
     wiki_subparser = wiki_parser.add_subparsers(dest="wiki_subcommand")
 
     wiki_download_subparser = wiki_subparser.add_parser(
@@ -74,3 +100,31 @@ def setup_wiki_command(main_subparser):
 
     wiki_download_subparser.set_defaults(func=process_wiki)
 
+
+def setup_output_command(main_subparser):
+    """Setup the top level output command line option."""
+
+    output_parser = main_subparser.add_parser(
+        "output", help="Command for performing output related tasks"
+    )
+    output_parser.set_defaults(func=output_parser.print_help)
+
+    output_parser.add_argument(
+        "wiki_cache_file", help="The path to the wiki data json file"
+    )
+
+    output_parser.add_argument(
+        "output_file", help="The path to the output html file"
+    )
+
+    output_parser.add_argument(
+        "--template_dir", required=False, default=DEFAULT_TEMPLATES_DIR,
+        help="Path to the directory holding the jinja templates",
+    )
+
+    output_parser.add_argument(
+        "--template_filename", required=False, default=FLINK_MAIN_PAGE_TEMPLATE,
+        help="Name of the flink main page template inside the template directory",
+    )
+
+    output_parser.set_defaults(func=process_output)
